+++ src/meta/mod.rs
@@ -36,17 +36,17 @@ use std::path::{Component, Path, PathBuf};
 pub struct Meta {
     pub name: Name,
     pub path: PathBuf,
-    pub permissions: Permissions,
-    pub date: Date,
-    pub owner: Owner,
+    pub permissions: Option<Permissions>,
+    pub date: Option<Date>,
+    pub owner: Option<Owner>,
     pub file_type: FileType,
-    pub size: Size,
+    pub size: Option<Size>,
     pub symlink: SymLink,
     pub indicator: Indicator,
-    pub inode: INode,
-    pub links: Links,
+    pub inode: Option<INode>,
+    pub links: Option<Links>,
     pub content: Option<Vec<Meta>>,
-    pub access_control: AccessControl,
+    pub access_control: Option<AccessControl>,
 }
 
 impl Meta {
@@ -169,17 +169,27 @@ impl Meta {
     }
 
     pub fn calculate_total_size(&mut self) {
+        if self.size.is_none() {
+            return;
+        }
+
         if let FileType::Directory { .. } = self.file_type {
             if let Some(metas) = &mut self.content {
-                let mut size_accumulated = self.size.get_bytes();
+                let mut size_accumulated = match &self.size {
+                    Some(size) => size.get_bytes(),
+                    None => 0,
+                };
                 for x in &mut metas.iter_mut() {
                     x.calculate_total_size();
-                    size_accumulated += x.size.get_bytes();
+                    size_accumulated += match &x.size {
+                        Some(size) => size.get_bytes(),
+                        None => 0,
+                    };
                 }
-                self.size = Size::new(size_accumulated);
+                self.size = Some(Size::new(size_accumulated));
             } else {
                 // possibility that 'depth' limited the recursion in 'recurse_into'
-                self.size = Size::new(Meta::calculate_total_file_size(&self.path));
+                self.size = Some(Size::new(Meta::calculate_total_file_size(&self.path)));
             }
         }
     }
@@ -225,6 +235,7 @@ impl Meta {
     pub fn from_path(path: &Path, dereference: bool) -> io::Result<Self> {
         let mut metadata = path.symlink_metadata()?;
         let mut symlink_meta = None;
+        let mut broken_link = false;
         if metadata.file_type().is_symlink() {
             match path.metadata() {
                 Ok(m) => {
@@ -238,7 +249,8 @@ impl Meta {
                     // This case, it is definitely a symlink or
                     // path.symlink_metadata would have errored out
                     if dereference {
-                        return Err(e);
+                        broken_link = true;
+                        eprintln!("lsd: {}: {}\n", path.to_str().unwrap_or(""), e);
                     }
                 }
             }
@@ -252,8 +264,6 @@ impl Meta {
         #[cfg(windows)]
         let (owner, permissions) = windows_utils::get_file_data(path)?;
 
-        let access_control = AccessControl::for_path(path);
-
         #[cfg(not(windows))]
         let file_type = FileType::new(&metadata, symlink_meta.as_ref(), &permissions);
 
@@ -261,16 +271,27 @@ impl Meta {
         let file_type = FileType::new(&metadata, symlink_meta.as_ref(), path);
 
         let name = Name::new(path, file_type);
-        let inode = INode::from(&metadata);
-        let links = Links::from(&metadata);
+
+        let (inode, links, size, date, owner, permissions, access_control) = match broken_link {
+            true => (None, None, None, None, None, None, None),
+            false => (
+                Some(INode::from(&metadata)),
+                Some(Links::from(&metadata)),
+                Some(Size::from(&metadata)),
+                Some(Date::from(&metadata)),
+                Some(owner),
+                Some(permissions),
+                Some(AccessControl::for_path(path)),
+            ),
+        };
 
         Ok(Self {
             inode,
             links,
             path: path.to_path_buf(),
             symlink: SymLink::from(path),
-            size: Size::from(&metadata),
-            date: Date::from(&metadata),
+            size,
+            date,
             indicator: Indicator::from(file_type),
             owner,
             permissions,
@@ -282,15 +303,61 @@ impl Meta {
     }
 }
 
-#[cfg(unix)]
 #[cfg(test)]
 mod tests {
     use super::Meta;
+    use std::fs::File;
+    use tempfile::tempdir;
 
+    #[cfg(unix)]
     #[test]
     fn test_from_path_path() {
         let dir = assert_fs::TempDir::new().unwrap();
         let meta = Meta::from_path(dir.path(), false).unwrap();
         assert_eq!(meta.path, dir.path())
+    }
+
+    #[test]
+    fn test_from_path() {
+        let tmp_dir = tempdir().expect("failed to create temp dir");
+
+        let path_a = tmp_dir.path().join("aaa.aa");
+        File::create(&path_a).expect("failed to create file");
+        let meta_a = Meta::from_path(&path_a, false).expect("failed to get meta");
+
+        let path_b = tmp_dir.path().join("bbb.bb");
+        let path_c = tmp_dir.path().join("ccc.cc");
+
+        #[cfg(unix)]
+        std::os::unix::fs::symlink(&path_c, &path_b).expect("failed to create broken symlink");
+
+        // this needs to be tested on Windows
+        // likely to fail because of permission issue
+        // see https://doc.rust-lang.org/std/os/windows/fs/fn.symlink_file.html
+        #[cfg(windows)]
+        std::os::windows::fs::symlink_file(&path_c, &path_b)
+            .expect("failed to create broken symlink");
+
+        let meta_b = Meta::from_path(&path_b, true).expect("failed to get meta");
+
+        assert!(
+            meta_a.inode.is_some()
+                && meta_a.links.is_some()
+                && meta_a.size.is_some()
+                && meta_a.date.is_some()
+                && meta_a.owner.is_some()
+                && meta_a.permissions.is_some()
+                && meta_a.access_control.is_some()
+        );
+
+        assert!(
+            meta_b.inode.is_none()
+                && meta_b.links.is_none()
+                && meta_b.size.is_none()
+                && meta_b.date.is_none()
+                && meta_b.owner.is_none()
+                && meta_b.permissions.is_none()
+                && meta_b.access_control.is_none()
+        );
     }
 }